add test for when getData card is enabled
